src/C/Env: Define C environment.
This is the 'heart' of the application, representing the current environment of
C variables.

We model this as a type environment, name -> addr map, and addr -> value
store (i.e. every name is associated with a memory location, which is thereby
associated with a value in the store).
This is a close-enough approximation to the model used by malloc/free/the C
runtime system, particularly for the purposes of basic points-to analysis.

Note that insert _updates_ variables without changing their underlying memory
location.
In some real environments, this might not be guaranteed!
Future versions may be smarter about re-definition / assignment, which is not
even implemented (technically).